ci(github-actions): Use `npm ci` instead of `npm install`
According to the documentation when the `package-lock.json` is committed to the repo,
`npm ci` is the preferred reliable way to install dependencies on CI,
since it will not attempt to modify the lock file.
It will also fail in case the dependencies in `package.json`
do not match with the ones in the lock file.

https://docs.npmjs.com/cli/ci