refactor: Switch `error/error.test.js`
There was a lot of crap in this one. The most crazy thing was that one test implied that all levels are writing a message, but it was infact just checking that the error handler was present in the dict (`errorHandler[k].length > 0` instead of `errorMap[k].length > 0`)
The previous commit moved one test case to a more suitable place.