fix: Handle whitespace in closing tags
by not treating them as part of the tag name.
We drop the additional whitespace as part of parsing
.
xmldom was already automatically closing tags that it couldn't find a matching end,
but the whitespace was no longer present.
Strange things started to happen if there were elements after the one with the additional white space.

Fixes #69