docs(readme): Improve wording and flow
to make it easier to digest.

I recently read some articles on how to structure a README and it became obvious that ours is currently hard to digest.
I think I was able to reorder the headlines so that the most important things for our (potential and existing) users are on the top of the page without a lot of scrolling.
There is still room for improvement, especially regarding splitting some docs into separate markdown files, but that is a slightly bigger topic.

https://dev.to/healeycodes/how-to-write-an-awesome-github-readme-2ldc
https://dev.to/jmfayard/how-to-write-a-good-readme-discuss-4hkl