chore: Split test and lint scripts
To fix #111 I think it's good enough to split test and lint into two different build steps.
Separate actions are not required since the github actions details view now shows what steps failed.

To be able to do that split, I needed to drop the execution of the `lint` script from the `test` script. The `start` script still runs both in watch mode, so it's still easy to run both when developing.