docs(readme): Add downloads per week badges
for both `xmldom` and `@xmldom/xmldom` since I keep opening both pages to check the progress of the transition.
I think having the badges at this spot will indicate when we will be able to remove this line from the readme.
It's worth mentioning that half of the downloads originates from deprecated versions of `xmldom`.