Merge pull request from GHSA-crh6-fp67-6883
* fix: Report errors if DOM is not well-formed

In case such a DOM would be created, the part that is not well formed will be transformed into text nodes, in which xml specific characters like `<` and `>` are encoded accordingly.

This change can break your code, if you relied on this behavior, e.g. multiple root elements in the past.
We consider it more important to align with the specs that we want to be aligned with,
considering the potential security issues that might derive from people not being aware of the difference in behavior.

Resolves https://github.com/xmldom/xmldom/security/advisories/GHSA-crh6-fp67-6883
https://dom.spec.whatwg.org/#concept-node-ensure-pre-insertion-validity

* fix: Prevent setting documentElement

if node is not inserted and add missing `Document.ownerDocument`.