fix: Simpler and more robust error handling
By passing `onError` option to `DOMParser` constructor,
it will be called for each kind of error that happens during parsing.
The main reason for joining the different handlers,
is that the current error levels make no sense,
and all should currently be treated as if they are errors.
If it throws the error will be wrapped into a ParseError to stop processing.
If it is not provided, xmldom continues to log to the console like before.
The previous `errorHandler` option has been marked deprecated.
If it is a function it will be used as `onError` if present (and `onError` is not passed), but logged as a warning.
If it is an object a `TypeError` will be thrown.

BREAKING-CHANGE: It is no longer possible to prevent the `fatalError` from throwing by overriding it. This ensures that no further processing takes place once a fatalError occurs.
Which is what the XML specifications demand.
This is why so many tests needed to be changed: Now they need to handle the throwing case instead of just collecting messages.