fix: Throw DOMException when calling removeChild with an invalid parameter
BREAKING-CHANGE: Previously it was possible (but not documented) to call `Node.removeChild` with any node in the tree,
and with certain exceptions, it would work. This is no longer the case: calling `Node.removeChild` this way will now throw a NotFoundError DOMException, as it is described by the specs.

https://dom.spec.whatwg.org/#concept-node-pre-remove
Fixes #135