fix(DOMParser): Align parseFromString errors with specs (#454)
- throw `TypeError` on invalid mimeType argument, it is now a mandatory
argument
- report `fatalError` if no `documentElement` was added, to ensure we
never return an invalid/incompletely initialized Document.
- report `fatalError` for non string source argument, instead of
`error`, but only for XML
- refactored MIME_TYPE to not contain functions

**fix: Simpler and more robust error handling**

There is a new option `onError` option that can be passed to the
`DOMParser` constructor:
The function that is invoked for every error that occurs during parsing.

If it is not provided, all errors are reported to `console.error`
and only `fatalError`s are thrown as a `ParseError`, which prevents any
further processing.
If `onError` throws, a `ParserError` is thrown, which prevents any
further processing.

The main reason for joining the different handlers, is that the current
error levels make no sense, and all should currently be treated as if
they are errors.

The previous `errorHandler` option has been marked deprecated.
If it is a function and `onError` is not passed it will be used as
`onError`, but is reported as a warning.
If it is an object a `TypeError` will be thrown as part of the
`DOMParser` constructor.

BREAKING CHANGE: It is no longer possible to prevent the `fatalError`
from throwing by setting `errorHandler`. This ensures that no further
processing takes place once a fatalError occurs.
Which is what the XML specifications demand.
This is why so many tests needed to be changed: Now they need to handle
the throwing case instead of just collecting messages.