refactor: merge redundant ']' condition at start of while loop
with the proposed fix `parseDoctypeInternalSubset` had two places that checked for the ending symbol and returned the internalSubSet string.
By moving it to the beginning of the while loop, we only need to have it once.

I also added tests for more cases ending with a space before the closing square bracket.