refactor: merge redundant ']' condition at start of while loop (#698)
with the fix added in #692 `parseDoctypeInternalSubset` has two places
that check for the ending symbol and returned the `internalSubSet` as a
string. By moving it to the beginning of the while loop, we only need to
check the condition once.

I also added tests for more cases ending with a space before the closing
square bracket.