fix: restore performance for large amount of child nodes
as part of https://github.com/xmldom/xmldom/pull/550
("drop redundant and broken `Element.appendChild` implementation")
the `_onUpdateChild` was not passed the added child,
so the `childNodes` were re-indexed from the `firstChild` each time a node was appended.

Now that the added child is being passed,
the performance in 0.9.0 is still not as good as in 0.8.*,
but this is because of the higher amount of checks being executed before adding a node.

fixes https://github.com/xmldom/xmldom/issues/748