fix: lower error level for unicode replacement character (#794)
I was misled by some test cases from xmltest.zip,
to think that xml containing the Unicode replacement character should be
treated as not well-formed. This is now fixed by lowering the error
level from `fatalError` which can not be ignored, to `warning` which can
be ignored.
It can also be treated as an exit condition by using
`onWarningStopParsing` as the error handler, which is recommended if you
do not need to be lax about the content being parsed.

fixes #790