fix: allow case-insensitive DOCTYPE in HTML
When introducing stricter parsing of the DOCTYPE and supporting more complex DTDs,
we have not been aware that the HTML spec allows any kind of casing but only certain values.

To make sure this change doesn't cause more harm,
xmldom will still parse anything beside DTD syntax in HTML,
and report warnings if there is anything unexpected in the systemId.
(One example of this is parsing XHTML as HTML.)

closes #817

https://html.spec.whatwg.org/multipage/syntax.html#the-doctype